+++ valhalla/midgard/tiles.h
@@ -161,7 +161,7 @@ public:
    *          extent, an error (-1) is returned.
    */
   int32_t TileId(const coord_t& c) const {
-    return TileId(c.y(), c.x());
+    return TileId(static_cast<float>(c.y()), static_cast<float>(c.x()));
   }
 
   /**
